refactor: OutputView 분리
