+++ src/main/kotlin/blackjack/controller/BlackjackController.kt
@@ -8,14 +8,16 @@ import blackjack.model.participant.Players
 import blackjack.state.State
 import blackjack.view.BettingAmountInputView
 import blackjack.view.IsAddCardInputView
-import blackjack.view.OutputView
 import blackjack.view.PlayersInputView
+import blackjack.view.ProgressView
+import blackjack.view.ResultView
 
 class BlackjackController(
     private val playersInputView: PlayersInputView = PlayersInputView(),
     private val bettingAmountInputView: BettingAmountInputView = BettingAmountInputView(),
     private val isAddCardInputView: IsAddCardInputView = IsAddCardInputView(),
-    private val outputView: OutputView = OutputView(),
+    private val progressView: ProgressView = ProgressView(),
+    private val resultView: ResultView = ResultView(),
 ) {
     private val deck: Deck = Deck()
     private val dealer: Dealer = Dealer(deck)
@@ -32,7 +34,7 @@ class BlackjackController(
 
     private fun setUpGame() {
         players = playersInputView.readPlayerNames(deck)
-        outputView.printInitCard(dealer, players)
+        progressView.printInitCard(dealer, players)
     }
 
     private fun setBettingAmount() {
@@ -50,12 +52,12 @@ class BlackjackController(
     private fun playerTurn(player: Player) {
         while (player.state !is State.Finished) {
             processPlayerStateDecision(player)
+            progressView.printPlayerCard(player)
         }
 
         if (player.isBust()) {
-            outputView.printBustMessage()
+            progressView.printBustMessage()
         }
-        outputView.printPlayerCard(player)
     }
 
     private fun processPlayerStateDecision(player: Player) {
@@ -70,13 +72,13 @@ class BlackjackController(
     private tailrec fun dealerTurn() {
         if (dealer.isAddCard()) {
             dealer.addCard()
-            outputView.printDealerAddCard()
+            progressView.printDealerAddCard()
             dealerTurn()
         }
     }
 
     private fun showResult() {
-        outputView.printCardResult(dealer, players)
-        outputView.printGameResult(judge.gameResult(dealer, players.gamePlayers))
+        resultView.printCardResult(dealer, players)
+        resultView.printGameResult(judge.gameResult(dealer, players.gamePlayers))
     }
 }
